blog_project/requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
boto3==1.17.53
botocore==1.20.53
cffi==1.14.5
Django==3.1.7
django-bootstrap4==2.3.1
django-braces==1.14.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-storages==1.11.1
Faker==6.5.0
jmespath==0.10.0
Pillow==8.1.1
pycparser==2.20
Pygments==2.8.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
s3transfer==0.3.7
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
